requirements.txt
binaryornot==0.3.0
blessings==1.6
click==3.3
cookiecutter==1.0.0
docopt==0.6.2
falcon==0.2
gnureadline==6.3.3
httplib2==0.9
ipython==3.0.0
Jinja2==2.7.3
MarkupSafe==0.23
mock==1.0.1
PIL==1.1.7
python-instagram==1.3.0
python-mimeparse==0.1.4
PyYAML==3.11
requests==2.5.3
simplejson==3.6.5
six==1.9.0
